refactor : 패키지 구조 변경 및 전 주차 과제 SOLID 원칙에 맞게 리팩토링
