+++ src/main/java/kr/hhplus/be/server/application/order/OrderCriteria.java
@@ -2,9 +2,11 @@ package kr.hhplus.be.server.application.order;
 
 import kr.hhplus.be.server.domain.payment.PaymentMethod;
 import kr.hhplus.be.server.interfaces.dto.OrderDTO;
-import kr.hhplus.be.server.support.utils.LockKeyPrefix;
+import kr.hhplus.be.server.support.aop.lock.LockKeyPrefix;
 
 import java.util.List;
+
+import static kr.hhplus.be.server.domain.payment.PaymentMethod.POINTS;
 
 public class OrderCriteria {
 	public record Cancel(
@@ -12,31 +14,36 @@ public class OrderCriteria {
 		Long orderId
 	) {
 		public static Cancel from(Long userId, Long orderId) {
-			return new Cancel(userId,orderId);
+			return new Cancel(userId, orderId);
 		}
-		public String getLockKey(){
+
+		public String getLockKey() {
 			return LockKeyPrefix.ORDER_CANCEL.createKey(orderId);
 		}
 	}
+
 	public record Request(
 		Long userId,
 		List<Request.Item> orderItems,
 		List<String> couponIds,
+		String requestId,
 		PaymentMethod paymentMethod
-	){
+	) {
 		public static Request from(OrderDTO.OrderRequest dto) {
 			List<Request.Item> items = dto.getProducts().stream()
 				.map(p -> new Request.Item(p.getProductId(), p.getQuantity()))
 				.toList();
 
-			return new Request(dto.getUserId(), items, dto.getCouponId(),PaymentMethod.POINTS);
+			return new Request(dto.getUserId(), items, dto.getCouponId(), dto.getRequestId(), POINTS);
 		}
-		public String getLockKey(){
+		public String getLockKey() {
 			return LockKeyPrefix.ORDER.createKey(userId);
 		}
-		public List<Long> productIds(){
+
+		public List<Long> productIds() {
 			return orderItems.stream().map(Item::productId).toList();
 		}
+
 		public record Item(
 			Long productId,
 			int quantity
