+++ src/main/java/kr/hhplus/be/server/support/aop/lock/DistributedLockAspect.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.infra.lock;
+package kr.hhplus.be.server.support.aop.lock;
 
 import kr.hhplus.be.server.support.common.exception.CustomException;
 import kr.hhplus.be.server.support.config.swagger.ErrorCode;
@@ -11,7 +11,6 @@ import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.reflect.MethodSignature;
 import org.redisson.api.RLock;
-import org.redisson.api.RedissonClient;
 import org.springframework.stereotype.Component;
 
 import java.lang.reflect.Method;
@@ -23,19 +22,21 @@ import java.util.concurrent.TimeUnit;
 @Slf4j
 public class DistributedLockAspect {
 
-	private final RedissonClient redissonClient;
+	//	private final RedissonClient redissonClient;
+	private final LockRepository lockRepository;
 	private final AopForTransaction aopForTransaction;
 
 	@Around("@annotation(redisLock)")
 	public Object lock(ProceedingJoinPoint joinPoint, RedisLock redisLock) throws Throwable {
+
+		long waitTime = redisLock.waitTime();
+		long leaseTime = redisLock.leaseTime();
 		Method method = ((MethodSignature) joinPoint.getSignature()).getMethod();
 		Object[] args = joinPoint.getArgs();
 
-		String lockKey = SpELParserHelper.parseExpression(redisLock.lockKey(), method, args);
-		long waitTime = redisLock.waitTime();
-		long leaseTime = redisLock.leaseTime();
-		log.info("### lock lockKey : {}", lockKey);
-		RLock lock = redissonClient.getLock(lockKey);
+		String parsedLockKey = SpELParserHelper.parseExpression(redisLock.lockKey(), method, args);
+		RLock lock = lockRepository.getLock(parsedLockKey);
+
 		boolean lockAcquired = lock.tryLock(waitTime, leaseTime, TimeUnit.MILLISECONDS);
 
 		if (!lockAcquired) {
