+++ src/main/java/kr/hhplus/be/server/support/aop/lock/LockRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.support.aop.lock;
+
+import org.redisson.api.RLock;
+
+public interface LockRepository {
+	RLock getLock(String lockKey);
+}
