+++ src/main/java/kr/hhplus/be/server/support/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.support.config.jpa;
+package kr.hhplus.be.server.support.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
